+++ packages/create-gatsby/package.json
@@ -16,6 +16,7 @@
     "execa": "^4.0.3",
     "fs-extra": "^9.0.1",
     "gatsby-core-utils": "^1.3.23",
+    "stream-filter": "^2.1.0",
     "terminal-link": "^2.1.1"
   },
   "files": [
