+++ packages/create-gatsby/src/types.d.ts
@@ -1,0 +1,1 @@
+declare module "stream-filter"
